# apps/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Triathlon API",
    "version": "0.2.0",
    "description": "One public HTTP interface for all Triathlon clients. Error responses share the Problem Details envelope with a stable domain code."
  },
  "components": {
    "securitySchemes": {
      "bearerKey": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/access-keys": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "ownerType": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "user"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "automation"
                        ]
                      }
                    ]
                  },
                  "automationId": {
                    "type": "string"
                  },
                  "scope": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "project"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "instance"
                        ]
                      }
                    ]
                  },
                  "projectId": {
                    "type": "string"
                  },
                  "expiresAt": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "key",
                    "secret"
                  ],
                  "properties": {
                    "key": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "ownerType",
                        "ownerUserId",
                        "automationId",
                        "scope",
                        "projectId",
                        "expiresAt",
                        "revokedAt",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "ownerType": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "user"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "automation"
                              ]
                            }
                          ]
                        },
                        "ownerUserId": {
                          "type": "string"
                        },
                        "automationId": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "scope": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "project"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "instance"
                              ]
                            }
                          ]
                        },
                        "projectId": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "expiresAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "revokedAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      }
                    },
                    "secret": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "ownerType",
                      "ownerUserId",
                      "automationId",
                      "scope",
                      "projectId",
                      "expiresAt",
                      "revokedAt",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "ownerType": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "user"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "automation"
                            ]
                          }
                        ]
                      },
                      "ownerUserId": {
                        "type": "string"
                      },
                      "automationId": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "scope": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "project"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "instance"
                            ]
                          }
                        ]
                      },
                      "projectId": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "expiresAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "revokedAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/access-keys/{keyId}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "keyId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/automations": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "ownerUserId",
                    "name",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "ownerUserId",
                      "name",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "ownerUserId": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/automations/{automationId}/transfer": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "toUserId"
                ],
                "properties": {
                  "toUserId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "automationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "ownerUserId",
                    "name",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "ownerUserId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "requestId",
                          "occurredAt",
                          "actorUserId",
                          "actorAutomationId",
                          "actorKeyId",
                          "clientHeader",
                          "operation",
                          "targetType",
                          "targetId",
                          "result",
                          "detail"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "requestId": {
                            "type": "string"
                          },
                          "occurredAt": {
                            "type": "string"
                          },
                          "actorUserId": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "actorAutomationId": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "actorKeyId": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "clientHeader": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "operation": {
                            "type": "string"
                          },
                          "targetType": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "targetId": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "result": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "ok"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "error"
                                ]
                              }
                            ]
                          },
                          "detail": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1"
    }
  ]
}